@@ -1,32 +1,32 @@
 name: fitnest_x
-description: A new Flutter project.
+description: Demo mobile app.
 
 publish_to: "none"
 
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^1.0.2
-  bloc: ^8.0.3
-  flutter_bloc: ^8.0.1
+  cupertino_icons: ^1.0.5
+  bloc: ^8.1.1
+  flutter_bloc: ^8.1.2
   get_it: ^7.2.0
-  google_fonts: ^3.0.0
-  equatable: ^2.0.3
-  intl: ^0.17.0
-  equations: ^4.0.0
-  camera: ^0.9.8+1
+  google_fonts: ^4.0.3
+  equatable: ^2.0.5
+  intl: ^0.18.0
+  equations: ^4.1.0
+  camera: ^0.10.3+2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^2.0.1
 
 flutter:
   uses-material-design: true
