@@ -94,7 +94,7 @@ class _ActivityProgressDayItem extends StatelessWidget {
               bottom: 0,
               child: Text(data.label,
                   textAlign: TextAlign.center,
-                  style: Theme.of(context).textTheme.subtitle1))
+                  style: Theme.of(context).textTheme.titleMedium))
         ],
       ),
     );
