@@ -51,14 +51,14 @@ class PopularMealCard extends StatelessWidget {
               Text(data.name,
                   maxLines: 1,
                   overflow: TextOverflow.ellipsis,
-                  style: textTheme.bodyText2?.copyWith(
+                  style: textTheme.bodyMedium?.copyWith(
                       fontWeight: FontWeight.bold, color: AppColors.black)),
               AppWhiteSpace.value3.vertical,
               Text(
                 subtitle,
                 maxLines: 1,
                 overflow: TextOverflow.ellipsis,
-                style: textTheme.subtitle1,
+                style: textTheme.titleMedium,
               )
             ],
           )),
