@@ -27,7 +27,7 @@ class DescriptionStepItem extends StatelessWidget {
           SizedBox(
             width: _kStepNumberWidth,
             child: Text('${index < 10 ? '0' : ''}$index',
-                style: textTheme.bodyText2?.copyWith(
+                style: textTheme.bodyMedium?.copyWith(
                     color: step.isActive ? AppColors.purple : AppColors.gray2)),
           ),
           AppWhiteSpace.value15.horizontal,
@@ -43,9 +43,9 @@ class DescriptionStepItem extends StatelessWidget {
               children: [
                 Text(step.title,
                     style:
-                        textTheme.bodyText2?.copyWith(color: AppColors.black)),
+                        textTheme.bodyMedium?.copyWith(color: AppColors.black)),
                 AppWhiteSpace.value3.vertical,
-                Text(step.description, style: textTheme.subtitle1),
+                Text(step.description, style: textTheme.titleMedium),
                 AppWhiteSpace.value30.vertical,
               ],
             ),
