@@ -50,10 +50,10 @@ class FindMealCard extends StatelessWidget {
         )),
         AppWhiteSpace.value10.vertical,
         Text(data.mealTime.title,
-            style: textTheme.bodyText2?.copyWith(
+            style: textTheme.bodyMedium?.copyWith(
                 color: AppColors.black, fontWeight: FontWeight.bold)),
         AppWhiteSpace.value3.vertical,
-        Text('${data.amount}+ Foods', style: textTheme.subtitle1),
+        Text('${data.amount}+ Foods', style: textTheme.titleMedium),
         AppWhiteSpace.value15.vertical,
         if (_cardColor == CardColor.blue)
           SecondaryButton.blue(
@@ -61,7 +61,7 @@ class FindMealCard extends StatelessWidget {
             padding: _kButtonPadding,
             onPressed: onPressed ?? () {},
             height: _kButtonHeight,
-            textStyle: textTheme.subtitle1
+            textStyle: textTheme.titleMedium
                 ?.copyWith(color: AppColors.white, fontWeight: FontWeight.bold),
           )
         else
@@ -70,7 +70,7 @@ class FindMealCard extends StatelessWidget {
             padding: _kButtonPadding,
             onPressed: onPressed ?? () {},
             height: _kButtonHeight,
-            textStyle: textTheme.subtitle1
+            textStyle: textTheme.titleMedium
                 ?.copyWith(color: AppColors.white, fontWeight: FontWeight.bold),
           )
       ]),
