@@ -22,9 +22,9 @@ class SleepEstimateCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final TextTheme textTheme = Theme.of(context).textTheme;
-    final digitTextStyle = textTheme.bodyText2
+    final digitTextStyle = textTheme.bodyMedium
         ?.copyWith(color: AppColors.black, fontWeight: FontWeight.bold);
-    final textStyle = textTheme.subtitle1?.copyWith(color: AppColors.black);
+    final textStyle = textTheme.titleMedium?.copyWith(color: AppColors.black);
 
     return Container(
       padding: _kCardPadding,
