@@ -21,7 +21,7 @@ class UpcomingWorkoutCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final color = data.color ?? AppColors.blue2;
-    final TextStyle? textStyle = Theme.of(context).textTheme.subtitle1;
+    final TextStyle? textStyle = Theme.of(context).textTheme.titleMedium;
 
     return AppCard(
       borderRadius: AppBorderRadius.value16.value,
