@@ -107,7 +107,7 @@ class _WorkoutChip extends StatelessWidget {
               child: Text(text,
                   style: Theme.of(context)
                       .textTheme
-                      .subtitle1
+                      .titleMedium
                       ?.copyWith(color: textColor)),
             ),
           ),
