@@ -33,13 +33,13 @@ class CongratulationsScreen extends StatelessWidget {
           ),
           AppWhiteSpace.value35.vertical,
           Text(_titleText,
-              textAlign: TextAlign.center, style: textTheme.headline4),
+              textAlign: TextAlign.center, style: textTheme.headlineMedium),
           AppWhiteSpace.value15.vertical,
           Text(_quoteText,
-              textAlign: TextAlign.center, style: textTheme.subtitle1),
+              textAlign: TextAlign.center, style: textTheme.titleMedium),
           AppWhiteSpace.value5.vertical,
           Text(_authorText,
-              textAlign: TextAlign.center, style: textTheme.subtitle1),
+              textAlign: TextAlign.center, style: textTheme.titleMedium),
           const Spacer(),
           PrimaryButton.blue(
               text: _backText,
